[Performance] Template Eager Load LCP (#817)
