+++ express/blocks/search-marquee/search-marquee.js
@@ -254,6 +254,8 @@ function decorateBackground(block) {
     const media = mediaRow.querySelector('picture img');
     if (media) {
       media.classList.add('backgroundimg');
+      media.loading = 'eager';
+      media.setAttribute('fetchpriority', 'high');
       const wrapper = block.parentElement;
       if (wrapper.classList.contains('search-marquee-wrapper')) {
         wrapper.prepend(media);
