+++ express/blocks/template-x/template-rendering.js
@@ -1,6 +1,9 @@
 /* eslint-disable no-underscore-dangle */
-import { createTag, getIconElement, getMetadata } from '../../scripts/utils.js';
-import BlockMediator from '../../scripts/block-mediator.min.js';
+import {
+  createTag,
+  getIconElement,
+  getMetadata,
+} from '../../scripts/utils.js';
 
 function containsVideo(pages) {
   return pages.some((page) => !!page?.rendition?.video?.thumbnail?.componentId);
@@ -344,7 +347,7 @@ function renderMediaWrapper(template, placeholders) {
   };
 }
 
-async function renderHoverWrapper(template, placeholders) {
+async function renderHoverWrapper(template, placeholders, isEligible) {
   const btnContainer = createTag('div', { class: 'button-container' });
 
   const {
@@ -354,22 +357,6 @@ async function renderHoverWrapper(template, placeholders) {
   btnContainer.append(mediaWrapper);
   btnContainer.addEventListener('mouseenter', enterHandler);
   btnContainer.addEventListener('mouseleave', leaveHandler);
-
-  let isEligible = document.body.dataset.device === 'desktop' || !['yes', 'true', 'Y', 'on'].includes(getMetadata('mobile-benchmark'));
-
-  if (!isEligible) {
-    const eligibility = BlockMediator.get('mobileBetaEligibility');
-    if (eligibility) {
-      isEligible = eligibility.deviceSupport;
-    } else {
-      isEligible = await new Promise((resolve) => {
-        const unsub = BlockMediator.subscribe('mobileBetaEligibility', (e) => {
-          resolve(e.newValue.deviceSupport);
-          unsub();
-        });
-      });
-    }
-  }
 
   if (isEligible) {
     const cta = renderCTA(placeholders, template.customLinks.branchUrl);
@@ -437,10 +424,10 @@ function renderStillWrapper(template, placeholders) {
   return stillWrapper;
 }
 
-export default async function renderTemplate(template, placeholders) {
+export default async function renderTemplate(template, placeholders, isEligible) {
   const tmpltEl = createTag('div');
   tmpltEl.append(renderStillWrapper(template, placeholders));
-  tmpltEl.append(await renderHoverWrapper(template, placeholders));
+  tmpltEl.append(await renderHoverWrapper(template, placeholders, isEligible));
 
   return tmpltEl;
 }
