+++ express/blocks/template-x/template-x.js
@@ -1,5 +1,3 @@
-/* eslint-disable import/named, import/extensions */
-
 import {
   createOptimizedPicture,
   createTag,
@@ -21,6 +19,7 @@ import { fetchTemplates, isValidTemplate, fetchTemplatesCategoryCount } from './
 import fetchAllTemplatesMetadata from '../../scripts/all-templates-metadata.js';
 import renderTemplate from './template-rendering.js';
 import isDarkOverlayReadable from '../../scripts/color-tools.js';
+import BlockMediator from '../../scripts/block-mediator.min.js';
 
 function wordStartsWithVowels(word) {
   return word.match('^[aieouâêîôûäëïöüàéèùœAIEOUÂÊÎÔÛÄËÏÖÜÀÉÈÙŒ].*');
@@ -39,9 +38,11 @@ function handlelize(str) {
     .toLowerCase(); // To lowercase
 }
 
-async function getTemplates(response, phs, fallbackMsg) {
+async function getTemplates(response, phs, fallbackMsg, isEligible) {
   const filtered = response.items.filter((item) => isValidTemplate(item));
-  const templates = await Promise.all(filtered.map((template) => renderTemplate(template, phs)));
+  const templates = await Promise.all(
+    filtered.map((template) => renderTemplate(template, phs, isEligible)),
+  );
   return {
     fallbackMsg,
     templates,
@@ -72,7 +73,7 @@ async function fetchAndRenderTemplates(props) {
   props.total = response.metadata.totalHits;
 
   // eslint-disable-next-line no-return-await
-  return await getTemplates(response, placeholders, fallbackMsg);
+  return await getTemplates(response, placeholders, fallbackMsg, props.isEligible);
 }
 
 async function processContentRow(block, props) {
@@ -1623,6 +1624,22 @@ export default async function decorate(block) {
   addTempWrapper(block, 'template-x');
 
   const props = constructProps(block);
+  // temporary for mobile beta eligibility
+  let isEligible = document.body.dataset.device === 'desktop' || !['yes', 'true', 'Y', 'on'].includes(getMetadata('mobile-benchmark'));
+  if (!isEligible) {
+    const eligibility = BlockMediator.get('mobileBetaEligibility');
+    if (eligibility) {
+      isEligible = eligibility.deviceSupport;
+    } else {
+      isEligible = await new Promise((resolve) => {
+        const unsub = BlockMediator.subscribe('mobileBetaEligibility', (e) => {
+          resolve(e.newValue.deviceSupport);
+          unsub();
+        });
+      });
+    }
+  }
+  props.isEligible = isEligible;
   block.innerHTML = '';
   await buildTemplateList(block, props, determineTemplateXType(props));
 }
